--- src/common/Wrapper/index.js
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const Wrapper = styled.div`
-    max-width: 1368px;
-  
-`;
-
-
