+++ src/features/PopularMovies/MovieDetails/index.js
@@ -3,7 +3,7 @@ import { useParams } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 
 import { Header } from "./Header";
-import { Wrapper } from "./styled";
+import { MovieAndPersonWrapper } from "../../../common/Wrapper/MovieAndPersonWrapper";
 import { BigMovieTile } from "../../Tiles/BigMovieTile";
 import PeopleTiles from "../../Tiles/PeopleTiles";
 import Error from "../../../common/Error";
@@ -29,11 +29,11 @@ export const MovieDetails = () => {
                 return (
                     <>
                         <Header />
-                        <Wrapper>
+                        <MovieAndPersonWrapper>
                             <BigMovieTile />
                             <PeopleTiles title="Cast" />
                             <PeopleTiles title="Crew" />
-                        </Wrapper>
+                        </MovieAndPersonWrapper>
                     </>
                 );
             case "error":
