+++ src/features/PopularMovies/index.js
@@ -5,7 +5,7 @@ import MovieTiles from "../Tiles/MovieTiles";
 import { Loader } from "../../common/Loader";
 import Error from "../../common/Error";
 import { NoResults } from "../../common/NoResults";
-import { Wrapper } from "./MovieDetails/styled";
+import { MovieAndPersonWrapper } from "../../common/Wrappers/MovieAndPersonWrapper";
 import { Pagination } from "./Pagination";
 import { useQueryParameter } from "../../useQueryParameter";
 import searchQueryParamName from "../../searchQueryParamName";
@@ -27,9 +27,9 @@ const PopularMoviesList = () => {
             case "success":
                 return (
                     <>
-                        <Wrapper>
+                        <MovieAndPersonWrapper>
                             <MovieTiles title="Popular movies" />
-                        </Wrapper>
+                        </MovieAndPersonWrapper>
                         <Pagination />
                     </>);
             case "error":
