+++ src/features/PopularPeople/PersonDetails/index.js
@@ -7,7 +7,7 @@ import MovieTiles from "../../Tiles/MovieTiles";
 import { Loader } from "../../../common/Loader";
 import Error from "../../../common/Error";
 import { NoResults } from "../../../common/NoResults";
-import { Wrapper } from "./styled";
+import { MovieAndPersonWrapper } from "../../../common/Wrappers/MovieAndPersonWrapper";
 import { fetchPersonDetails, selectStatus } from "./personDetailsSlice";
 
 const PersonDetails = () => {
@@ -26,11 +26,11 @@ const PersonDetails = () => {
             case "success":
                 return (
                     <>
-                        <Wrapper>
+                        <MovieAndPersonWrapper>
                             <PersonTile />
                             <MovieTiles title="cast"/>
                             <MovieTiles title="crew" />
-                        </Wrapper>
+                        </MovieAndPersonWrapper>
                     </>
                 );
             case "error":
